Refactored client
