+++ concept/thing/impl/RelationImpl.java
@@ -19,7 +19,7 @@
 
 package grakn.client.concept.thing.impl;
 
-import grakn.client.Transaction;
+import grakn.client.GraknClient;
 import grakn.client.concept.Concept;
 import grakn.client.concept.ConceptId;
 import grakn.client.concept.thing.Attribute;
@@ -54,7 +54,7 @@ public class RelationImpl {
      */
     public static class Remote extends ThingImpl.Local.Remote<Relation, RelationType> implements Relation.Remote {
 
-        public Remote(Transaction tx, ConceptId id) {
+        public Remote(GraknClient.Transaction tx, ConceptId id) {
             super(tx, id);
         }
