+++ test/behaviour/graql/GraqlSteps.java
@@ -20,8 +20,7 @@
 package grakn.client.test.behaviour.graql;
 
 import com.google.common.collect.Iterators;
-import grakn.client.Session;
-import grakn.client.Transaction;
+import grakn.client.GraknClient;
 import grakn.client.answer.Answer;
 import grakn.client.answer.ConceptMap;
 import grakn.client.answer.Explanation;
@@ -63,8 +62,8 @@ import static org.junit.Assert.assertTrue;
 
 public class GraqlSteps {
 
-    private static Session session = null;
-    private static Transaction tx = null;
+    private static GraknClient.Session session = null;
+    private static GraknClient.Transaction tx = null;
 
     private static List<ConceptMap> answers;
     HashMap<String, UniquenessCheck> identifierChecks = new HashMap<>();
