+++ test/deployment/src/test/java/application/MavenApplicationTest.java
@@ -1,8 +1,6 @@
 package application;
 
 import grakn.client.GraknClient;
-import grakn.client.Session;
-import grakn.client.Transaction;
 import grakn.client.answer.ConceptMap;
 import graql.lang.Graql;
 import org.junit.Test;
@@ -16,8 +14,8 @@ public class MavenApplicationTest {
     @Test
     public void testImport() {
         GraknClient client = GraknClient.open("localhost:48555");
-        Session session = client.session("grakn");
-        Transaction tx = session.transaction().write();
+        GraknClient.Session session = client.session("grakn");
+        GraknClient.Transaction tx = session.transaction().write();
         List<ConceptMap> answers = tx.execute(Graql.match(var("t").sub("thing")).get());
         tx.close();
         session.close();
