+++ test/integration/answer/AnswerIT.java
@@ -20,8 +20,6 @@
 package grakn.client.test.integration.answer;
 
 import grakn.client.GraknClient;
-import grakn.client.Session;
-import grakn.client.Transaction;
 import grakn.client.answer.ConceptMap;
 import grakn.client.answer.Explanation;
 import grakn.client.test.setup.GraknProperties;
@@ -68,8 +66,8 @@ public class AnswerIT {
 
     @Test
     public void testExplanation() {
-        Session session = client.session("test_rules");
-        Transaction tx = session.transaction().write();
+        GraknClient.Session session = client.session("test_rules");
+        GraknClient.Transaction tx = session.transaction().write();
         tx.execute(Graql.parse(" define\n" +
                 "                    object sub entity, plays owned, plays owner;\n" +
                 "                    ownership sub relation, relates owned, relates owner;\n" +
