+++ test/integration/tracing/TracingIT.java
@@ -22,8 +22,6 @@ package grakn.client.test.integration.tracing;
 import grabl.tracing.client.GrablTracingThreadStatic;
 import grabl.tracing.client.GrablTracingThreadStatic.ThreadContext;
 import grakn.client.GraknClient;
-import grakn.client.Session;
-import grakn.client.Transaction;
 import grakn.client.answer.ConceptMap;
 import grakn.client.test.setup.GraknProperties;
 import grakn.client.test.setup.GraknSetup;
@@ -68,8 +66,8 @@ public class TracingIT {
     @Test
     public void testWithTracing() {
         try (ThreadContext ignored = contextOnThread("tracker", 1)) {
-            try (Session session = client.session("test_tracing")) {
-                try (Transaction tx = session.transaction().write()) {
+            try (GraknClient.Session session = client.session("test_tracing")) {
+                try (GraknClient.Transaction tx = session.transaction().write()) {
                     tx.execute(Graql.parse("define\n" +
                             "name sub attribute, datatype string;\n" +
                             "person sub entity, has name;").asDefine());
